chore: rename src/factory/ -> src/b20_factory/ per nit
