+++ crates/common/precompiles/src/b20_factory/storage.rs
