+++ crates/common/precompiles/src/b20_factory/variant.rs
